requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cloudinary==1.21.1
Django==3.0
django-bootstrap4==2.1.1
djangorestframework==3.11.0
importlib-metadata==1.6.1
linecache2==1.0.0
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.9
zipp==3.1.0
